splice_agent/requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Data access
duckdb>=0.9.0
pyarrow>=14.0.0
openpyxl>=3.1.0  # For Excel support

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# FastAPI service
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.6.0

# Optional: Jupyter support
jupyter>=1.0.0
ipykernel>=6.25.0
